+++ teilleistung3/flexlater/app/models/Student.java
@@ -11,14 +11,14 @@ public class Student extends Model {
 
 	// Attribute
 	@Id
-	public Long studentId;
+	public Integer studentId;
 	@Required
 	public String name, firstName, major;
 	
 
 	// Methoden
 
-	public static Finder<Long, Student> find = new Finder (Long.class, Student.class);
+	public static Finder<Integer, Student> find = new Finder (Integer.class, Student.class);
 
 	public static List<Student> all() {
 		return find.all();
@@ -28,7 +28,7 @@ public class Student extends Model {
 		student.save();
 	}
 
-	public static void delete(Long studentId) {
+	public static void delete(Integer studentId) {
 		find.ref(studentId).delete();
 	}
